re-implemented lucene search
